pyproject.toml
[project]
name = "lumenparser"
version = "0.1.0"
description = "A RAG toolkit for unstructured document noise and contextual loss in chunking"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bm25s>=0.2.14",
    "chromadb>=1.5.0",
    "docling>=2.73.1",
    "einops>=0.8.2",
    "google-genai>=1.64.0",
    "langchain-chroma>=1.1.0",
    "langchain-huggingface>=1.2.0",
    "mteb>=2.7.30",
    "pymupdf-layout>=1.27.1",
    "pymupdf4llm>=0.3.4",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.1",
    "transformers>=4.57.6",
]
